Setter functions added and additional spacing
The setter functions here can be easily removed but I decided to add them since we may want to change Date to a class with private member variables in the future. It also gives us more lines. If you decide to remove them be sure to change the Transaction::setDate function as well.